Import `Id` struct from the `bxcan` crate
Its a superset of the current `Id` with minor API changes.
Fully device independent still.
