+++ src/lib.rs
@@ -3,99 +3,19 @@
 
 pub mod blocking;
 
-/// Standard 11bit Identifier (0..=0x7FF)
-#[derive(Debug, Copy, Clone, Eq, PartialEq)]
-pub struct StandardId(u16);
+mod id;
 
-impl StandardId {
-    /// Creates a new standard identifier.
-    pub fn new(id: u16) -> Result<StandardId, ()> {
-        if id <= 0x7FF {
-            Ok(StandardId(id))
-        } else {
-            Err(())
-        }
-    }
-}
-
-impl core::convert::From<StandardId> for u16 {
-    fn from(id: StandardId) -> u16 {
-        id.0
-    }
-}
-
-impl core::convert::From<StandardId> for u32 {
-    fn from(id: StandardId) -> u32 {
-        id.0 as u32
-    }
-}
-
-impl ExtendedId {
-    /// Creates a new extended identifier.
-    pub fn new(id: u32) -> Result<ExtendedId, ()> {
-        if id <= 0x1FFF_FFFF {
-            Ok(ExtendedId(id))
-        } else {
-            Err(())
-        }
-    }
-}
-
-impl core::convert::From<ExtendedId> for u32 {
-    fn from(id: ExtendedId) -> u32 {
-        id.0
-    }
-}
-
-/// Extended 29bit Identifier (0..=0x1FFF_FFFF)
-#[derive(Debug, Copy, Clone, Eq, PartialEq)]
-pub struct ExtendedId(u32);
-
-/// CAN Identifier
-///
-/// The variants are wrapped in newtypes so they can only be costructed with valid values.
-#[derive(Debug, Copy, Clone, Eq, PartialEq)]
-pub enum Id {
-    /// Standard 11bit Identifier (0..=0x7FF)
-    Standard(StandardId),
-
-    /// Extended 29bit Identifier (0..=0x1FFF_FFFF)
-    Extended(ExtendedId),
-}
-
-impl Id {
-    /// Creates a new standard identifier.
-    pub fn new_standard(id: u16) -> Result<Id, ()> {
-        Ok(StandardId::new(id)?.into())
-    }
-
-    /// Creates a new extended identifier.
-    pub fn new_extended(id: u32) -> Result<Id, ()> {
-        Ok(ExtendedId::new(id)?.into())
-    }
-}
-
-impl core::convert::From<StandardId> for Id {
-    fn from(id: StandardId) -> Id {
-        Id::Standard(id)
-    }
-}
-
-impl core::convert::From<ExtendedId> for Id {
-    fn from(id: ExtendedId) -> Id {
-        Id::Extended(id)
-    }
-}
+pub use id::*;
 
 /// A CAN2.0 Frame
 pub trait Frame: Sized {
     /// Creates a new frame.
     /// Returns an error when the data slice is too long.
-    fn new(id: Id, data: &[u8]) -> Result<Self, ()>;
+    fn new(id: impl Into<Id>, data: &[u8]) -> Result<Self, ()>;
 
     /// Creates a new remote frame (RTR bit set).
     /// Returns an error when the data length code (DLC) is not valid.
-    fn new_remote(id: Id, dlc: usize) -> Result<Self, ()>;
+    fn new_remote(id: impl Into<Id>, dlc: usize) -> Result<Self, ()>;
 
     /// Returns true if this frame is a extended frame.
     fn is_extended(&self) -> bool;
